requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-ckeditor==5.9.0
django-import-export==2.2.0
django-js-asset==1.2.2
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
